requirements.txt
pymongo==3.7.2
flask==1.0.2
pyJWT==1.4.0
Werkzeug==0.14
Flask-SQLAlchemy==2.3.2